Adding findAll with limit to EntityDao

+++ src/main/java/net/dontdrinkandroot/persistence/dao/JpaGenericDao.java
@@ -17,7 +17,11 @@
  */
 package net.dontdrinkandroot.persistence.dao;
 
-import java.util.List;
+import net.dontdrinkandroot.persistence.entity.Entity;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
@@ -27,13 +31,7 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Root;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-import net.dontdrinkandroot.persistence.entity.Entity;
+import java.util.List;
 
 
 /**
@@ -182,7 +180,7 @@ public class JpaGenericDao implements GenericDao
 
 		criteriaQuery.select(count);
 
-		return this.findSingle(criteriaQuery).longValue();
+		return this.findSingle(criteriaQuery);
 	}
 
 	@Override
